+++ src/components/__tests__/Auth.test.tsx
@@ -6,9 +6,14 @@ import { State } from "opds-web-client/lib/state";
 import { AuthState } from "opds-web-client/lib/reducers/auth";
 import fetchMock from "fetch-mock-jest";
 import { AuthCredentials } from "opds-web-client/lib/interfaces";
+import userEvent from "@testing-library/user-event";
 
 const stateWithAuth: State = merge<State>(fixtures.initialState, {
-  auth: fixtures.unauthenticatedAuthState
+  auth: fixtures.unauthenticatedAuthState,
+  loans: {
+    url: "/loans-url",
+    books: []
+  }
 });
 
 afterEach(() => {
@@ -32,7 +37,7 @@ test("shows overlay on unauthenticated request", async () => {
     }
   );
   // overlay should not be shown yet
-  const modal = utils.getByLabelText("Sign In");
+  const modal = utils.getByLabelText("Sign In Form");
   expect(modal).toBeInTheDocument();
   expect(modal).toHaveStyle("display: none");
 
@@ -69,7 +74,7 @@ test("shows overlay when showForm is true", async () => {
     }
   );
   // overlay should be shown
-  const modal = utils.getByLabelText("Sign In");
+  const modal = utils.getByLabelText("Sign In Form");
   expect(modal).toBeInTheDocument();
   expect(modal).toHaveStyle("display: block");
 });
@@ -84,11 +89,56 @@ test("renders auth form provided in authPlugin", async () => {
     }
   );
   // overlay should be shown
-  const modal = utils.getByLabelText("Sign In");
+  const modal = utils.getByLabelText("Sign In Form");
   expect(modal).toBeInTheDocument();
   expect(modal).toHaveStyle("display: block");
   // should include the BasicAuthForm
-  expect(utils.getByLabelText("Pin")).toBeInTheDocument();
+  expect(utils.getByLabelText("Pin input")).toBeInTheDocument();
+});
+
+const authStateWithTwoProviders: AuthState = {
+  showForm: true,
+  callback: jest.fn(),
+  cancel: jest.fn(),
+  credentials: null,
+  title: "form",
+  error: null,
+  attemptedProvider: null,
+  providers: [fixtures.basicAuthProvider, fixtures.samlAuthProvider]
+};
+const stateWithTwoProviders: State = merge<State>(fixtures.initialState, {
+  auth: authStateWithTwoProviders
+});
+
+test("renders select when multiple providers present", async () => {
+  const utils = render(
+    <Auth>
+      <div>children</div>
+    </Auth>,
+    {
+      initialState: stateWithTwoProviders
+    }
+  );
+
+  const select = utils.getByRole("combobox", { name: "Login Method" });
+  expect(select).toBeInTheDocument();
+
+  // should have two options
+  expect(
+    utils.getByRole("option", { name: "Library Barcode" })
+  ).toBeInTheDocument();
+  expect(utils.getByRole("option", { name: "SAML IdP" })).toBeInTheDocument();
+
+  // should default to basic auth (first)
+  expect(
+    utils.getByRole("textbox", { name: "Barcode input" })
+  ).toBeInTheDocument();
+  expect(utils.getByLabelText("Pin input")).toBeInTheDocument();
+
+  // should be able to change
+  userEvent.selectOptions(select, fixtures.samlAuthProvider.id);
+
+  expect(await utils.findByText("Login with SAML IdP")).toBeInTheDocument();
 });
 
 test("displays message when no auth provider configured", async () => {
@@ -120,12 +170,14 @@ test("displays message when no auth provider configured", async () => {
     }
   );
   // overlay should be shown
-  const modal = utils.getByLabelText("Sign In");
+  const modal = await utils.findByLabelText("Sign In Form");
   expect(modal).toBeInTheDocument();
   expect(modal).toHaveStyle("display: block");
 
   expect(
-    utils.getByText("Basic auth provider is missing.")
+    utils.getByText(
+      "There is no Auth Plugin configured for the selected Auth Provider."
+    )
   ).toBeInTheDocument();
 });
 
@@ -146,7 +198,7 @@ test("attempts to get credentials from cookies", async () => {
 test("attempts to save auth credentials", async () => {
   const credentials: AuthCredentials = {
     credentials: "auth-cred",
-    provider: fixtures.basicAuthType
+    provider: fixtures.basicAuthId
   };
   const getCredentialsSpy = jest.spyOn(fetcher, "getAuthCredentials");
   getCredentialsSpy.mockReturnValue(credentials);
@@ -164,7 +216,5 @@ test("attempts to save auth credentials", async () => {
   expect(setCredentialsSpy).toHaveBeenCalledTimes(1);
   expect(setCredentialsSpy).toHaveBeenCalledWith(credentials);
   expect(fetchLoansSpy).toHaveBeenCalledTimes(1);
-  expect(fetchLoansSpy).toHaveBeenLastCalledWith(
-    "http://test-cm.com/catalogUrl/loans"
-  );
+  expect(fetchLoansSpy).toHaveBeenLastCalledWith("/loans-url");
 });